fix(gameplay): round immediate perk health effects

+++ src/crimson/perks/impl/infernal_contract.py
@@ -1,5 +1,6 @@
 from __future__ import annotations
 
+from ...math_parity import f32
 from ..ids import PerkId
 from ..runtime.apply_context import PerkApplyCtx
 from ..runtime.hook_types import PerkHooks
@@ -12,7 +13,7 @@ def apply_infernal_contract(ctx: PerkApplyCtx) -> None:
         ctx.perk_state.choices_dirty = True
     for player in ctx.players:
         if player.health > 0.0:
-            player.health = 0.1
+            player.health = f32(0.1)
 
 
 HOOKS = PerkHooks(
